Create README.md
## 📝 About CalculateIT

CalculateIT was created as a foundational project to build a solid understanding of basic web development concepts.  
It served as an initial step 🚀 to practice and strengthen skills in HTML, CSS, and JavaScript through building a functional and interactive calculator.  
This project helped me grasp core principles like  event handling 🎯, and responsive design 📱.